Use a long to store number of zones advanced to avoid overflow

+++ Source/Castro.cpp
@@ -37,7 +37,7 @@ BCRec        Castro::phys_bc;
 int          Castro::NUM_STATE     = -1;
 int          Castro::NUM_GROW      = -1;
 
-int          Castro::num_zones_advanced = 0;
+long         Castro::num_zones_advanced = 0;
 
 Real         Castro::frac_change   = 1.e200;
